Moved configuration related plugin to appropriate package
