+++ src/main/groovy/org/shipkit/internal/gradle/ReleaseNotesPlugin.java
@@ -6,6 +6,7 @@ import org.gradle.api.Project;
 import org.shipkit.gradle.ReleaseConfiguration;
 import org.shipkit.gradle.ReleaseNotesFetcherTask;
 import org.shipkit.gradle.UpdateReleaseNotesTask;
+import org.shipkit.internal.gradle.configuration.ReleaseConfigurationPlugin;
 import org.shipkit.internal.gradle.git.GitPlugin;
 import org.shipkit.internal.gradle.util.TaskMaker;
 import org.shipkit.internal.version.VersionInfo;
