+++ src/main/groovy/org/shipkit/internal/gradle/configuration/ReleaseConfigurationPlugin.java
@@ -1,10 +1,11 @@
-package org.shipkit.internal.gradle;
+package org.shipkit.internal.gradle.configuration;
 
 import org.gradle.api.Action;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 import org.gradle.api.plugins.ObjectConfigurationAction;
 import org.shipkit.gradle.ReleaseConfiguration;
+import org.shipkit.internal.gradle.VersioningPlugin;
 import org.shipkit.internal.gradle.init.InitConfigFileTask;
 import org.shipkit.internal.gradle.init.InitPlugin;
 import org.shipkit.internal.gradle.util.TaskMaker;
