+++ src/main/groovy/org/shipkit/internal/gradle/release/PublicationsComparatorPlugin.java
@@ -11,8 +11,8 @@ import org.shipkit.internal.comparison.DownloadPreviousReleaseArtifactsTask;
 import org.shipkit.internal.comparison.PublicationsComparatorTask;
 import org.shipkit.internal.comparison.artifact.DefaultArtifactUrlResolver;
 import org.shipkit.internal.comparison.artifact.DefaultArtifactUrlResolverFactory;
-import org.shipkit.internal.gradle.ReleaseConfigurationPlugin;
 import org.shipkit.internal.gradle.configuration.DeferredConfiguration;
+import org.shipkit.internal.gradle.configuration.ReleaseConfigurationPlugin;
 import org.shipkit.internal.gradle.java.JavaLibraryPlugin;
 import org.shipkit.internal.gradle.java.JavaPublishPlugin;
 import org.shipkit.internal.gradle.util.TaskMaker;
