+++ src/test/groovy/org/shipkit/internal/gradle/VersioningPluginTest.groovy
@@ -1,6 +1,7 @@
 package org.shipkit.internal.gradle
 
 import org.shipkit.gradle.BumpVersionFileTask
+import org.shipkit.internal.gradle.configuration.ReleaseConfigurationPlugin
 import org.shipkit.internal.gradle.git.GitCommitTask
 import org.shipkit.internal.gradle.git.GitPlugin
 import org.shipkit.internal.gradle.init.InitVersioningTask
