+++ src/test/groovy/org/shipkit/internal/gradle/configuration/ReleaseConfigurationPluginTest.groovy
@@ -1,4 +1,4 @@
-package org.shipkit.internal.gradle
+package org.shipkit.internal.gradle.configuration
 
 import org.gradle.api.Project
 import org.gradle.testfixtures.ProjectBuilder
