+++ src/test/groovy/testutil/PluginSpecification.groovy
@@ -5,7 +5,7 @@ import org.gradle.testfixtures.ProjectBuilder
 import org.junit.Rule
 import org.junit.rules.TemporaryFolder
 import org.shipkit.gradle.ReleaseConfiguration
-import org.shipkit.internal.gradle.ReleaseConfigurationPlugin
+import org.shipkit.internal.gradle.configuration.ReleaseConfigurationPlugin
 import org.shipkit.internal.notes.util.IOUtil
 import spock.lang.Specification
